tradeai-v2/backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Cache & Queue
redis==5.0.1
celery==5.3.4

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Data Processing
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2

# ML/AI Libraries
scikit-learn==1.3.2
tensorflow==2.15.0
prophet==1.1.5

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Production
sentry-sdk[fastapi]==1.38.0